reimplementando cosas que saque por error

+++ Assets/Scritp/Personajes/Genericos/IDamager.cs
@@ -1,4 +1,4 @@
 public interface IDamager
 {
-     float Damage ();
+     int Damage ();
 }
